+++ src/shared/infra/http/index.ts
@@ -31,19 +31,35 @@ import bodyParser from 'body-parser'
 const app = express()
 const appName = env.APP_NAME
 const appPort = env.APP_PORT
-const appTest = env.TEST_APP_URL
 
 const rateLimit = new RateLimiter({ limit: 50, per: 'minutes' })
 
 app.use(
   cors({
-    allowedHeaders: ['Content-Type', 'Authorization'],
+    allowedHeaders: [
+      'Content-Type',
+      'Authorization',
+      'Accept',
+      'Accept-Language',
+      'Accept-Encoding',
+      'Origin',
+      'Referer',
+      'Cookie',
+      'Sec-Fetch-Dest',
+      'Sec-Fetch-Mode',
+      'Sec-Fetch-Site',
+      'Pragma',
+      'Cache-Control',
+      'Access-Control-Allow-Origin',
+    ],
     methods: ['GET', 'POST', 'PUT', 'PATCH', 'DELETE'],
     origin: [
       'https://www.ognare.com.br',
       'https://ognare.com.br',
-      appTest ?? '',
+      'http://localhost',
+      'http://localhost:3000',
     ],
+    preflightContinue: false,
     credentials: true,
   }),
 )
@@ -97,8 +113,8 @@ if (env.NODE_ENV !== 'dev') {
 app.use((err: Error, req: Request, res: Response, next: NextFunction) => {
   if (err instanceof AppError) {
     return res.status(err.statusCode).json({
-      errorTitle: err.title,
-      errorMessage: err.message,
+      ok: false,
+      error: err,
     })
   }
 
@@ -106,16 +122,24 @@ app.use((err: Error, req: Request, res: Response, next: NextFunction) => {
     if (env.NODE_ENV === 'dev') console.log(err)
 
     return res.status(400).json({
-      errorTitle: 'Informações inválidas',
-      errorMessage: 'Verifique as informações fornecidas e tente novamente',
+      ok: false,
+      error: {
+        title: 'Informações inválidas',
+        message: 'Verifique as informações fornecidas e tente novamente',
+        statusCode: 400,
+      },
     })
   }
 
   if (err instanceof MulterError && err.code === 'LIMIT_FILE_SIZE') {
     return res.status(400).json({
-      errorTitle: 'Imagem maior que 2 mb',
-      errorMessage:
-        'O limite de tamanho de imagens aceito é 2 mb nos planos free.',
+      ok: false,
+      error: {
+        title: 'Imagem maior que 2 mb',
+        message:
+          'O limite de tamanho de imagens aceito é 2 mb nos planos free.',
+        statusCode: 400,
+      },
     })
   }
 
@@ -123,8 +147,12 @@ app.use((err: Error, req: Request, res: Response, next: NextFunction) => {
     if (env.NODE_ENV === 'dev') console.log(err)
 
     res.status(500).json({
-      errorTitle: 'Internal error',
-      errorMessage: 'Internal error',
+      ok: false,
+      error: {
+        title: 'Internal error',
+        message: 'Internal error',
+        statusCode: 500,
+      },
     })
   }
 })